Exception handling - delete

+++ src/main/java/com/educandoweb/course/services/UserService.java
@@ -3,8 +3,10 @@ package com.educandoweb.course.services;
 import java.util.List;
 import java.util.Optional;
 
+import com.educandoweb.course.resources.exceptions.DatabaseException;
 import com.educandoweb.course.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.educandoweb.course.entities.User;
@@ -29,8 +31,16 @@ public class UserService {
 		return repository.save(obj);
 	}
 
-	public void delete(Long id){
-		repository.deleteById(id);
+	public void delete(Long id) {
+		try {
+			if (repository.existsById(id)) {
+				repository.deleteById(id);
+			} else {
+				throw new ResourceNotFoundException(id);
+			}
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
 
 	public User update( Long id, User obj){
